backend/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
pydantic>=2.5.0
pydantic-settings>=2.1.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
ruff>=0.1.0
pre-commit>=3.6.0
types-python-jose>=3.3.0
types-passlib>=1.7.8
